+++ src/SlobSteak.Application/ServiceCollectionExtensions.cs
@@ -1,6 +1,7 @@
 using Microsoft.Extensions.DependencyInjection;
 using SlobSteak.Application.Assessments;
 using SlobSteak.Application.Identity;
+using SlobSteak.Application.Map;
 using SlobSteak.Application.Projects;
 using SlobSteak.Application.Stakeholders;
 
@@ -33,6 +34,7 @@ public static class ServiceCollectionExtensions
         services.AddScoped<GetStakeholderService>();
         services.AddScoped<UpsertStakeholderAssessmentService>();
         services.AddScoped<GetStakeholderAssessmentsQuery>();
+        services.AddScoped<StakeholderMapQuery>();
 
         return services;
     }